cli: Implement `rbs_rails server` command
To make RBS generation faster to extract signatures from the Ruby on
Rails application, this provides the rbs_rails LSP server.  It generates
signatures on demand if the IDE lets the rbs_rails LSP server know the
changes to Ruby files.

It preloads the Rails application on start and reloads it when the
source code is changed. It improves the performance of signature
generation rather than running the rbs_rails command

+++ test/test_helper.rb
@@ -3,11 +3,13 @@ require 'minitest/autorun'
 require 'minitest/hooks'
 
 require 'pathname'
+require 'bundler'
 require 'rbs_rails'
 
 TEST_SIG_DIR = Pathname(File.expand_path('./app/sig/rbs_rails', __dir__))
 
 def clean_test_signatures
+  return unless TEST_SIG_DIR.exist?
   File.delete(*TEST_SIG_DIR.glob('**/*.rbs'))
 end
 
@@ -24,12 +26,40 @@ def expectations_dir
   Pathname(__dir__).join('expectations')
 end
 
-def setup!
+def setup!(rbs_rails: true)
   dir = app_dir
 
   Bundler.with_unbundled_env do
     sh!('bundle', 'install', chdir: dir)
     sh!('bin/rake', 'db:create', 'db:schema:load', chdir: dir)
-    sh!('bin/rake', 'rbs_rails:all', '--trace', chdir: dir)
+    sh!('bin/rake', 'rbs_rails:all', '--trace', chdir: dir) if rbs_rails
+  end
+end
+
+# Constant stubbing capabilities for tests
+module ConstStubbable
+  def stub_const(const_name, new_value)
+    @stub_const ||= {}
+
+    const_name = const_name.to_sym
+    if Object.const_defined?(const_name, false)
+      @stub_const[const_name] = Object.const_get(const_name)
+      Object.send(:remove_const, const_name)
+    else
+      @stub_const[const_name] = nil
+    end
+
+    Object.const_set(const_name, new_value)
+  end
+
+  def clean_const_stubs
+    return unless @stub_const
+
+    @stub_const.each do |const_name, original|
+      Object.send(:remove_const, const_name) if Object.const_defined?(const_name, false)
+      Object.const_set(const_name, original) if original
+    end
+
+    @stub_const.clear
   end
 end
